Move BoardState definition from `types.h` to `position.h`
